refactor(dicom-web-api): renamed from dicomWeb
